shared corruption tester script i had laying around.
Had this script laying around for a few years now, needs PERL to run but by looking at it.
It seems to run a 7zip test command on each archive (user set extention) then write failed archives filenames into a text file.

Might have to look into editing this to make ComicRack Lists instead so its loadable in Comicrack more easily.

Need to figure out a neat way to use this with dupes.

When loaded into Comicrack it could flag files with a field allowing you to add that to the dupe manager script in ComicRack maybe. That makes sense.